Appropriately move postNewFood() fetch request which now calls renderFoods() [which moved to response-handler]
