+++ lib/event-listeners/foods-event-listeners.js
@@ -1,7 +1,6 @@
 const $ = require('jquery')
-const postNewFood = require('../response-handlers/foods-response-handler').postNewFood
+const postNewFood = require('../fetch-requests/foods-fetch-requests').postNewFood
 const missingFoodField = require('../response-handlers/foods-response-handler').missingFoodField
-const renderNewFood = require('../response-handlers/foods-response-handler').renderNewFood
 const deleteFoodRecord = require('../fetch-requests/foods-fetch-requests').deleteFoodRecord
 
 const newFoodSubmit = () => {
@@ -14,7 +13,6 @@ const newFoodSubmit = () => {
     if (name === "" || calories === "") return missingFoodField(name, calories)
 
     postNewFood(name, calories)
-    renderNewFood(name, calories)
   })
 }
