+++ lib/fetch-requests/foods-fetch-requests.js
@@ -2,7 +2,8 @@ const $ = require('jquery')
 const url = `https://protected-basin-11627.herokuapp.com/`
 import {handleResponse} from '../response-handlers/initial-response-handler'
 const removeFood = require('../response-handlers/foods-response-handler').removeFood
-
+const renderFoods = require('../response-handlers/foods-response-handler').renderFoods
+const renderNewFood = require('../response-handlers/foods-response-handler').renderNewFood
 
 const getAllFoods = () => {
   fetch(`${url}/api/v1/foods`)
@@ -11,28 +12,30 @@ const getAllFoods = () => {
     .catch(error => console.error({ error }))
 }
 
-function renderFoods(foods) {
-  foods.forEach(food => {
-    $('#foods-table').append(`
-      <tr data-food-id="${food.id}">
-        <td>${food.name}</td>
-        <td>${food.calories}</td>
-        <td class="delete-btns">
-          <img data-delete-btn-id="${food.id}" src="http://icons.iconarchive.com/icons/hopstarter/sleek-xp-basic/24/Delete-icon.png" alt="Delete Button">
-        </td>
-      </tr>
-      `)
-  })
-}
-
 function deleteFoodRecord(id) {
   fetch(`${url}/api/v1/foods/${id}`, { method: 'DELETE' })
     .then(removeFood(id))
     .catch(error => console.alert(error))
 }
 
+function postNewFood(name, calories) {
+  let newFood = { food: {'name': name, 'calories': calories}}
+
+  let postOptions = {
+    method: "POST",
+    body: JSON.stringify(newFood),
+    headers: {
+      'Content-Type': 'application/json',
+    },
+  }
+  fetch(`${url}/api/v1/foods`, postOptions)
+    .then(response => handleResponse(response))
+    .then(food => renderNewFood(food))
+}
+
 module.exports = {
   getAllFoods,
   deleteFoodRecord,
   renderFoods,
+  postNewFood,
 }
