+++ lib/response-handlers/foods-response-handler.js
@@ -2,18 +2,19 @@ const $ = require('jquery')
 const url = `https://protected-basin-11627.herokuapp.com/`
 
 const handleResponse = require('../response-handlers/initial-response-handler')
-const renderFoods = require('../fetch-requests/foods-fetch-requests').renderFoods
 
-const postNewFood = (name, calories) => {
-  let newFood = { food: {'name': name, 'calories': calories}}
-
-  let postOptions = {
-    method: "POST",
-    body: JSON.stringify(newFood),
-    headers: {'Content-Type': 'application/json'},
-  }
-  fetch(`${url}/api/v1/foods`, postOptions)
-  .then(response => handleResponse(response))
+function renderFoods(foods) {
+  foods.forEach(food => {
+    $('#foods-table').append(`
+      <tr data-food-id="${food.id}">
+        <td>${food.name}</td>
+        <td>${food.calories}</td>
+        <td class="delete-btns">
+          <img data-delete-btn-id="${food.id}" src="http://icons.iconarchive.com/icons/hopstarter/sleek-xp-basic/24/Delete-icon.png" alt="Delete Button">
+        </td>
+      </tr>
+      `)
+  })
 }
 
 function removeFood(id) {
@@ -25,14 +26,14 @@ const missingFoodField = (name, calories) => {
   if (calories === "") $('#missing-calories-alert').append("Please enter a calorie amount")
 }
 
-const renderNewFood = (name, calories) => {
-  let foodInArray = [{ "name": name, "calories": calories }]
+const renderNewFood = (food) => {
+  let foodInArray = [food]
 
   renderFoods(foodInArray)
 }
 
 module.exports = {
-  postNewFood,
+  renderFoods,
   removeFood,
   missingFoodField,
   renderNewFood,
